deployment/requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.1.0
boto3==1.33.13
botocore==1.33.13
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
dnspython==2.4.2
fastjsonschema==2.19.0
idna==3.6
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-notebook-parser==0.1.4
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
packaging==23.2
pandocfilters==1.5.0
platformdirs==4.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
s3transfer==0.8.2
six==1.16.0
soupsieve==2.5
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.0
urllib3==2.0.7
webencodings==0.5.1
